+++ CHANGELOG.md
@@ -16,6 +16,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Changed
   - Split the CoNLL-2009/mate format into a Gold and a System version
+    - The format previously known as `conll2009` is now `conll2009_gold`
+    - We have a new `conll2009_sys` that uses predicted columns instead of gold ones
 
 ## [0.5.1] - 2017-04-03
 [0.5.1]: https://github.com/LoicGrobol/ginger/compare/v0.5.0...v0.5.1