+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 [Unreleased]: https://github.com/LoicGrobol/ginger/compare/v0.6.0...HEAD
+### Fixed
+  - Format guessing now plays nice with CoNLL-2009
+  - Issue proper errors on wrong number of columns for CoNLL-2009
 
 ## [0.6.0] - 2017-04-03
 [0.6.0]: https://github.com/LoicGrobol/ginger/compare/v0.5.1...v0.6.0